Fragment upload logic separately
