+++ v2/src/components/App.tsx
@@ -1,22 +1,11 @@
-import React, { useState } from 'react';
-import { TransitionGroup } from 'react-transition-group';
-import { v4 as uuidv4 } from 'uuid';
+import React from 'react';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
+import Chip from '@mui/material/Chip';
 // import Typography from '@mui/material/Typography';
 // import Link from '@mui/material/Link';
-import Chip from '@mui/material/Chip';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import Collapse from '@mui/material/Collapse';
 
-import FileCard from './FileCard';
-import UploadButton from './UploadButton';
-
-interface IdentifiableFile {
-  file: File;
-  uuid: string;
-}
+import Upload from './Upload';
 
 // function Copyright() {
 //   return (
@@ -32,27 +21,6 @@ interface IdentifiableFile {
 // }
 
 function App() {
-  const [files, setFiles] = useState<IdentifiableFile[]>([]);
-
-  const handleAppendFile = (file: File) => {
-    const identifiableFile = { file, uuid: uuidv4() };
-    setFiles((prev) => [...prev, identifiableFile]);
-  };
-
-  const handleCancel = (uuid: string) => {
-    setFiles((prev) => [...prev.filter((file: IdentifiableFile) => file.uuid !== uuid)]);
-  };
-
-  const handleRetry = (uuid: string) => {
-    // Assign a new UUID to force re-rendering
-    setFiles((prev) => prev.map<IdentifiableFile>((file: IdentifiableFile) => (file.uuid !== uuid
-      ? file
-      : {
-        ...file,
-        uuid: uuidv4(),
-      })));
-  };
-
   return (
     <Container maxWidth='sm'>
       <Box sx={{ my: 4 }}>
@@ -79,31 +47,7 @@ function App() {
           Create React App example with TypeScript
         </Typography> */}
         {/* <Copyright /> */}
-        <List sx={{
-          width: '100%',
-          // maxHeight: 200,
-          maxHeight: 800,
-          position: 'relative',
-          overflow: 'auto',
-          '& ul': { padding: 0 },
-        }}
-        >
-          <TransitionGroup>
-            {files.map(({ file, uuid }) => (
-              <Collapse key={uuid}>
-                <ListItem sx={{ p: 0, m: 0 }} divider>
-                  <FileCard
-                    file={file}
-                    uuid={uuid}
-                    onRetry={() => handleRetry(uuid)}
-                    onCancel={() => handleCancel(uuid)}
-                  />
-                </ListItem>
-              </Collapse>
-            ))}
-          </TransitionGroup>
-        </List>
-        <UploadButton onClick={handleAppendFile} />
+        <Upload />
       </Box>
     </Container>
   );
